Approved -need to update paths the same way for other files

+++ public/js/profile.js
@@ -1,23 +1,48 @@
-// remove movie from collection
-const removeMovieHandler = async (event) => {
+// display all movies belong to user on page load
+const loadAllMoviesHandler = async (event) => {
   event.preventDefault();
 
-  const response = await fetch('/user/:id' , {
-    method: 'DELETE',
-  });
+  const userID = req.sessions.userId;
 
-  if (response.ok) {
-    document.location.replace('/user/:id');
-  } else {
-    alert('Failed to remove Movie');
+  if (userID) {
+    const response = await fetch('/user/' + userID, {
+      method: 'GET',
+      body: userID,
+    });
+
+    if (response.ok) {
+      document.location.replace('/user/' + userID);
+    } else {
+      res.status(500).json({ message: 'Oops, something went wrong.' });
+    }
   }
 };
 
-// add event listener to activate
+// remove movie from collection
+const removeMovieHandler = async (event) => {
+  event.preventDefault();
+  // make variable that holds the movie id of the movie that is being deleted by the click of a button
+  const doomedId = document.querySelector('#movie-id').value.trim();
+  // make variable that holds the user id of the user who wants to delete the movie
+  const userID = req.sessions.userId;
+  //send that id through a fetch delete request
+  if (doomedId && userID) {
+    const response = await fetch('/user/' + userID, {
+      method: 'DELETE',
+      body: doomedId,
+    });
+
+    if (response.ok) {
+      document.location.replace('/user/' + userID);
+    } else {
+      alert('Failed to remove Movie');
+    }
+  }
+};
+
+// add event listeners to activate
 document
   .querySelector('.removeMovie-button')
-  .addEventListener('click', removeMovieHandler);
+  .addEventListener('submit', removeMovieHandler);
 
-
-const router = require('express').Router();
-const { User } = require('../models');
+document.addEventListener('load', loadAllMoviesHandler);
